Create lambda_shared package for lambdas
