+++ tests/ci/autoscale_runners_lambda/app.py
@@ -2,25 +2,19 @@
 
 """The lambda to decrease/increase ASG desired capacity based on current queue"""
 
-import json
 import logging
-import time
 from dataclasses import dataclass
 from pprint import pformat
 from typing import Any, List, Literal, Optional, Tuple
 
 import boto3  # type: ignore
-import requests  # type: ignore
 
-RUNNER_TYPE_LABELS = [
-    "builder",
-    "func-tester",
-    "func-tester-aarch64",
-    "fuzzer-unit-tester",
-    "stress-tester",
-    "style-checker",
-    "style-checker-aarch64",
-]
+from lambda_shared import (
+    CHException,
+    ClickHouseHelper,
+    RUNNER_TYPE_LABELS,
+    get_parameter_from_ssm,
+)
 
 ### Update comment on the change ###
 # 4 HOUR - is a balance to get the most precise values
@@ -74,61 +68,7 @@ def get_scales(runner_type: str) -> Tuple[int, int]:
     return scale_down, scale_up
 
 
-### VENDORING
-def get_parameter_from_ssm(name, decrypt=True, client=None):
-    if not client:
-        client = boto3.client("ssm", region_name="us-east-1")
-    return client.get_parameter(Name=name, WithDecryption=decrypt)["Parameter"]["Value"]
-
-
-class CHException(Exception):
-    pass
-
-
-class ClickHouseHelper:
-    def __init__(
-        self,
-        url: Optional[str] = None,
-        user: Optional[str] = None,
-        password: Optional[str] = None,
-    ):
-        self.url = url
-        self.auth = {}
-        if user:
-            self.auth["X-ClickHouse-User"] = user
-        if password:
-            self.auth["X-ClickHouse-Key"] = password
-
-    def _select_and_get_json_each_row(self, db, query):
-        params = {
-            "database": db,
-            "query": query,
-            "default_format": "JSONEachRow",
-        }
-        for i in range(5):
-            response = None
-            try:
-                response = requests.get(self.url, params=params, headers=self.auth)
-                response.raise_for_status()
-                return response.text
-            except Exception as ex:
-                logging.warning("Cannot fetch data with exception %s", str(ex))
-                if response:
-                    logging.warning("Reponse text %s", response.text)
-                time.sleep(0.1 * i)
-
-        raise CHException("Cannot fetch data from clickhouse")
-
-    def select_json_each_row(self, db, query):
-        text = self._select_and_get_json_each_row(db, query)
-        result = []
-        for line in text.split("\n"):
-            if line:
-                result.append(json.loads(line))
-        return result
-
-
-CH_CLIENT = ClickHouseHelper(get_parameter_from_ssm("clickhouse-test-stat-url"), "play")
+CH_CLIENT = None  # type: Optional[ClickHouseHelper]
 
 
 def set_capacity(
@@ -222,6 +162,9 @@ def main(dry_run: bool = True) -> None:
     asg_client = boto3.client("autoscaling")
     try:
         global CH_CLIENT
+        CH_CLIENT = CH_CLIENT or ClickHouseHelper(
+            get_parameter_from_ssm("clickhouse-test-stat-url"), "play"
+        )
         queues = CH_CLIENT.select_json_each_row("default", QUEUE_QUERY)
     except CHException as ex:
         logging.exception(
