+++ tests/ci/autoscale_runners_lambda/requirements.txt
@@ -1,1 +1,1 @@
-requests<2.30
+../lambda_shared_package
