+++ tests/ci/lambda_shared_package/.gitignore
@@ -1,0 +1,2 @@
+build
+*.egg-info
