+++ tests/ci/team_keys_lambda/build_and_deploy_archive.sh
@@ -3,6 +3,7 @@ set -xeo pipefail
 
 WORKDIR=$(dirname "$0")
 WORKDIR=$(readlink -f "${WORKDIR}")
+DIR_NAME=$(basename "$WORKDIR")
 cd "$WORKDIR"
 
 # Do not deploy the lambda to AWS
@@ -12,9 +13,8 @@ PY_VERSION=${PY_VERSION:-3.10}
 PY_EXEC="python${PY_VERSION}"
 # Image to build the lambda zip package
 DOCKER_IMAGE="python:${PY_VERSION}-slim"
-LAMBDA_NAME=$(basename "$WORKDIR")
 # Rename the_lambda_name directory to the-lambda-name lambda in AWS
-LAMBDA_NAME=${LAMBDA_NAME//_/-}
+LAMBDA_NAME=${DIR_NAME//_/-}
 # The name of directory with lambda code
 PACKAGE=lambda-package
 rm -rf "$PACKAGE" "$PACKAGE".zip
@@ -23,8 +23,9 @@ cp app.py "$PACKAGE"
 if [ -f requirements.txt ]; then
   VENV=lambda-venv
   rm -rf "$VENV" lambda-package.zip
-  docker run --rm --user="${UID}" --volume="${WORKDIR}:/lambda" --workdir="/lambda" "${DOCKER_IMAGE}" \
-    /bin/bash -c "
+  docker run --rm --user="${UID}" -e HOME=/tmp \
+    --volume="${WORKDIR}/..:/ci" --workdir="/ci/${DIR_NAME}" "${DOCKER_IMAGE}" \
+    /bin/bash -exc "
       '$PY_EXEC' -m venv '$VENV' &&
       source '$VENV/bin/activate' &&
       pip install -r requirements.txt
@@ -35,5 +36,5 @@ fi
 ( cd "$PACKAGE" && zip -9 -r ../"$PACKAGE".zip . )
 
 if [ -z "$DRY_RUN" ]; then
-  aws lambda update-function-code --function-name "$LAMBDA_NAME" --zip-file fileb://"$PACKAGE".zip
+  aws lambda update-function-code --function-name "$LAMBDA_NAME" --zip-file fileb://"$WORKDIR/$PACKAGE".zip
 fi
